fixed provider tests
